Re-enable HYSTERESIS
Ran into a scenario—HR drops just a bit under the threshold and stays there. In that case I'd prefer the fan to not lower the speed. My current hypothesis is that the work and the cooling are in a sort of equilibrium at that point and if I lower the fan speed, it might nudge the HR up again, leading to the fan sped thrashing.